fix bug

+++ lidar_camera_calibration/lidar_camera_calib_impl.cpp
@@ -173,7 +173,7 @@ void Calibration::Projection(const Eigen::Matrix3d &rot, const Eigen::Vector3d &
     cv::Mat rgb_image_project = cv::Mat::zeros(camera_->h(), camera_->w(), CV_8UC3);
     for (size_t i = 0; i < pts_3d.size(); ++i) {
         Eigen::Vector2d point_2d = camera_->project(rot * pts_3d[i] + tran);
-        if (!camera_->valid(point_2d)) {
+        if (!camera_->valid(point_2d.cast<int>())) {
             continue;
         }
 
@@ -301,7 +301,7 @@ void Calibration::BuildVPnp(const Eigen::Matrix3d &rot, const Eigen::Vector3d &t
     PointCloud::Ptr lidar_edge_cloud_2d(new PointCloud);
     for (size_t i = 0; i < lidar_edge_ps_2d.size(); i++) {
         lidar_edge_ps_2d[i] = camera_->project(rot * lidar_edge_ps_3d[i] + tran);
-        if (!camera_->valid(lidar_edge_ps_2d[i]))
+        if (!camera_->valid(lidar_edge_ps_2d[i].cast<int>()))
             continue;
         double x = lidar_edge_ps_2d[i].x();
         double y = lidar_edge_ps_2d[i].y();
